Refactoring ID3v2FrameReader/ID3v2TagReader
(preparation for chapters parsing)
